move reg code to registrationpy
